[DevToolsGardener] Fix `git cl format --js --full` causing an error
`RunGitDiffCmd` with `--name-only` returns the files separated
by a new line as a string and then `FindFilesToFormat` splits it
to convert it to a files list.

Previously, we were calling `keys()` on this list and it
was throwing an error. Simply returning the file names
should be enough.

See the failure in:
https://logs.chromium.org/logs/devtools-frontend/buildbucket/cr-buildbucket/8703934325829680961/+/u/Scripted_rolls/Update_Browser_Protocol_deps/Run_Browser_Protocol_script/stdout

Bug: 386840832
Change-Id: I7bf54089726a66867c9126b1c398e8e57cc6eb26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/6943433
Reviewed-by: Scott Lee <ddoman@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>