[git_cl] Remove the auto cleanup from disabled new auth
This is in preparation for removing the old auto config code.  New
auth has been rolled out stably for a long time now, and in any case
the auto config/cleanup code isn't super reliable (hence why we're
removing it).

We remove this final callsite before deleting all of the relevant
code, and leave a message of assistance in case anyone needs it.

Bug: 446999231
Change-Id: Ia13e5d75a89a4ddfb178ee9e4fe2090596493218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/6976601
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Gavin Mak <gavinmak@google.com>